use unsigned int instead of long

+++ src/parser/pipeline/exit.h
@@ -6,7 +6,9 @@
 #include "_enum.h"
 
 PARSER_PIPELINE_STATUS ParserTypingExit(
-    struct ParserTypingBuffer *prefix, struct ParserTypingBuffer *suffix, long hold_offset)
+    struct ParserTypingBuffer *prefix,
+    struct ParserTypingBuffer *suffix,
+    unsigned int hold_offset)
 {
   if (*(prefix->tail - 1) == 4)
     HLIB_TERMINAL_PRINTF_ENDL("\nexit", TERMINAL_HIGHLIGHT), exit(0);
